+++ Sources/PeripheryKit/Analyzer/Analyzer.swift
@@ -33,7 +33,6 @@ public final class Analyzer {
         DefaultConstructorReferenceBuilder.self,
         ComplexPropertyAccessorReferenceBuilder.self,
         EnumCaseReferenceBuilder.self,
-        AssociatedTypeTypeAliasReferenceBuilder.self,
 
         UnusedParameterRetainer.self,
         AssetReferenceRetainer.self,
@@ -52,7 +51,8 @@ public final class Analyzer {
         AssignOnlyPropertyReferenceEliminator.self,
 
         DeclarationMarker.self,
-        RedundantProtocolMarker.self
+        RedundantProtocolMarker.self,
+        LetShorthandPropertyReferenceRetainer.self
     ]
 
     private let graph: SourceGraph