+++ Tests/Fixtures/RetentionFixtures/testRetainsParamUsedInOverriddenMethod.swift
@@ -4,7 +4,7 @@ public class FixtureClass101Base {
     // No overrides, unused.
     public func func1(param: String) {}
 
-    // Overriden, used only in base.
+    // Overridden, used only in base.
     public func func2(param: String) { print(param) }
 
     // Used in override.
@@ -13,8 +13,17 @@ public class FixtureClass101Base {
     // Used in deeply nested override.
     public func func4(param: String) {}
 
+    // Same name with different types, used to validate accuracy.
+    public func func4(param: Int) {}
+
     // Overridden, unused.
     public func func5(param: String) {}
+
+    // Overridden, declared in subclass extension.
+    public func func6(param: String) {}
+
+    // Overridden in multiple subclass branches.
+    public func func7(param1: String, param2: String) {}
 }
 
 public class FixtureClass101Subclass1: FixtureClass101Base {
@@ -23,6 +32,12 @@ public class FixtureClass101Subclass1: FixtureClass101Base {
     public override func func3(param: String) {
         print(param)
     }
+
+    public override func func4(param: Int) {}
+
+    public override func func7(param1: String, param2: String) {
+        print(param1)
+    }
 }
 
 public class FixtureClass101Subclass2: FixtureClass101Subclass1 {
@@ -30,7 +45,15 @@ public class FixtureClass101Subclass2: FixtureClass101Subclass1 {
         print(param)
     }
 
+    public override func func4(param: Int) {}
+
     public override func func5(param: String) {}
+    
+    public override func func7(param1: String, param2: String) {}
+}
+
+public class FixtureClass101Subclass3: FixtureClass101Base {
+    public override func func7(param1: String, param2: String) {}
 }
 
 public class FixtureClass101InheritForeignBase: NSObject {