+++ Tests/Fixtures/UnusedParameterFixtures/testLocalVariableAssignment.swift
@@ -1,10 +1,11 @@
 import Foundation
 
 class SyntaxFixture4 {
-    func myFunc(param1: String, param2: String, param3: String?, param4: String?) {
+    func myFunc(param1: String, param2: String, param3: String?, param4: String?, param5: String?) {
         let _ = param1.description
         let _ = param2
         if let _ = param3 {}
         if let param4 {}
+        guard let param5 else { return }
     }
 }